config.py
from dotenv import load_dotenv
import os

load_dotenv()

# TODO: Load your API keys from environment variables
OPENWEATHER_API_KEY = os.getenv("OPENWEATHER_API_KEY")
WEATHERAPI_API_KEY = os.getenv("WEATHERAPI_API_KEY")
