requirements.txt
aiohttp
python-dotenv
pandas
pytest
